+++ app/models/foi_attachment/maskable.rb
@@ -4,8 +4,6 @@ module FoiAttachment::Maskable
   extend ActiveSupport::Concern
 
   included do
-    delegate :apply_masks, to: :info_request
-
     scope :masked, -> { where.not(masked_at: nil) }
     scope :unmasked, -> { where(masked_at: nil) }
   end
@@ -21,7 +19,7 @@ module FoiAttachment::Maskable
   def mask
     return if masking_failed?
 
-    body = apply_masks(unmasked_body, content_type)
+    body = apply_masks(unmasked_body, content_type, redacted_attribute: :body)
     body = sanitise_html(body) if content_type == 'text/html'
     update(body: body, masked_at: Time.zone.now)
 
@@ -40,6 +38,13 @@ module FoiAttachment::Maskable
 
       sibling.mask
     end
+  end
+
+  def apply_masks(text, content_type = 'text/plain', redacted_attribute:, redactable: self)
+    info_request.apply_masks(
+      text, content_type,
+      redactable: redactable, redacted_attribute: redacted_attribute
+    )
   end
 
   private